Added support for Mapbox radiuses parameter
To add this during the setup of the L.Routing.mapbox constructor add include the radiuses parameter.
e.g. L.Routing.mapbox(mapbox_token, { radiuses: 3000 })

Mapbox Radiuses - Maximum distance in meters that each coordinate is allowed to move when snapped to a nearby road segment.

Allows the ability to generate an error for an unreachable road segment. rather than creating a straight line.