feat: Use custom esbuild script for bundling files
In the first place, this allows for easier configuration and plugins.
Another nice benefit (which I'll have to try to completely verify) is that the bundle output supports different formats.
Hopefully this would decrease bundle size for bundlers such as esbuild/webpack etc. since the large deps do not get automatically included anymore